subsequent commit

+++ src/main/java/com/example/loan24/dto/request/PaymentRequest.java
@@ -3,7 +3,8 @@ package com.example.loan24.dto.request;
 import com.example.loan24.data.model.enumClass.PaymentType;
 import lombok.*;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
+
 
 @Builder
 @Getter
@@ -12,6 +13,6 @@ import java.math.BigInteger;
 @NoArgsConstructor
 public class PaymentRequest {
     private String email;
-    private BigInteger amount;
+    private BigDecimal amount;
     private PaymentType paymentType;
 }
